Add a new rule to generate the custom grpc interfaces
Soong was reading the content of the symbolic link instead of the
target of it. Thus instead of reading the content of protoc-gen-custom_grpc it
was reading the content of protoc-gen-custom_grpc.py, which was a
symbolic link to protoc-gen-custom_grpc.

Bug: 245747653
Test: manual
Ignore-AOSP-First: cherry-picked from aosp
Merged-In: I37c0af2ffa6803ad1114fccbff4fdcf6914c54c1

Change-Id: I37c0af2ffa6803ad1114fccbff4fdcf6914c54c1